display app title bar immediately
